simulation/requirements.txt
fastapi
uvicorn
jinja2

requests
paho-mqtt